requirements.txt
pydantic-ai
fastapi
uvicorn
trafilatura
beautifulsoup4
python-dotenv
openai
